pyproject.toml
[tool.poetry]
name = "soothsayer"
version = "0.1.0"
description = ""
authors = ["Nao-Y1996 <60519909+Nao-Y1996@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
google-api-python-client = "^2.159.0"
pydantic = "^2.10.5"
gradio = "^5.12.0"
flatlib = "0.2.3"
pyswisseph = "2.08.00-1"
google-cloud-aiplatform = "^1.79.0"
python-dotenv = "^1.0.1"
pydantic-ai = "^0.0.20"
google-generativeai = "^0.8.4"
grpcio = "1.60.1"
redis = "^5.2.1"
pyopenjtalk = "^0.4.0"
simpleaudio = "^1.0.4"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"




sqlalchemy = "^2.0.38"
pytestarch = "^3.1.1"
alembic = "^1.14.1"
psycopg2-binary = "^2.9.10"
mypy = "^1.15.0"
black = "^25.1.0"
[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"

isort = "^6.0.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
astrology-server = "src.astrology_result_server:main"
livechat-store = "src.store_livechat:main"